requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
gevent==20.6.2
tensorflow==2.3.0
tensorflow-estimator==2.3.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.18.5
h5py==2.10.0
Pillow==7.2.0
gunicorn==19.6.0